Add native Windows test jobs to CI matrix
This expands the CI test matrix in the main testing workflow to
test on both Ubuntu and Windows, instead of just Ubuntu.

It does not attempt to merge in the Cygwin workflow at this time,
which may or may not end up being useful to do in the future.

The new Windows test jobs all fail currently: the runs fail as a
result of various tests being consistently unable to pass but not
yet being marked xfail (or skip). It should be feasible to mark
these xfail with informative messages, but this commit doesn't
include that.