Update git-source citation in Reference.set_object
Reference.set_object explains its handling of oldbinsha by quoting
a comment from refs.c in the Git source code. However, that comment
now appears in refs/files-backend.c in that codebase. This updates
the reference so readers can look it up and find the comment in its
surrounding context.

The commit to the git project's source code that moved the code
that includes that comment is:

https://github.com/git/git/commit/7bd9bcf372d4c03bb7034346d72ae1318e2d0742