Add missing info in Submodule.remove docstring
This rewords and adds some missing information to the docstring of
Submodule.remove, for the "method" parameter, discussed in #1712.
It uses the second suggestion presented in that issue, formatted
in the style of the surrounding docstring (an 88-column wrap).

It also does some other rewording in that docstring, for clarity,
at the end of that same section (on the "method" parameter), and a
small punctuation fix in the section about the "force" parameter.