Don't pre-clone submodules on CI, at least for now
This has actions/checkout no longer automatically clone submodules
in the CI test workflows. This change is for the purpose of
reproducing #1713, to allow the forthcoming fix for it to be
tested.

However, continuing to rely on init-tests-after-clone.sh to get the
submodules would serve as a kind of regression testing for #1713.
So it is unclear at this time if and when this change should be
undone.