Remove now-unused import in top-level __init__.py
The inspect module was only used to dynamically generate __all__,
which is statically written since c862845. Although it includes
everything it originally had for compatibility (8edc53b), this did
not include modules, since the comprehension used to generate it
omitted those (which is what it needed the inspect module for).
So it was not, and is not, listed in __all__, and can be removed.