Never modify sys.path
This removes the logic that, under some (most) circumstances when
not using a version from PyPI, would insert the location of the
gitdb git-submodule near the front of sys.path.

(As noted in #1717, the specific way this was being done was not
causing the git-submodule's version of gitdb to actually be used.
But it was still modifying sys.path, which this now prevents.)

The installation test, which had verified the insertion into
sys.path, is modified accordingly, except that for now the check
that the very first sys.path entry is undisturbed is kept in place.