Document how to use vendored dependencies
This adds a subsection to the end of the installation instructions
in the readme to explain how to cause the versions of the gitdb
and/or smmap dependencies that are vendored as submodules of the
GitPython repository to be used, instead of the PyPI versions, in
the infrequent case that this is desired.

This goes along with he removal of the logic that conditionally
modified sys.path, since that logic was intended to facilitate this
(and at one time had). The approach now documented in the readme
uses editable installs and does not involve modifying sys.path.

Because the need for this is uncommon, it may end up being moved
entirely into documentation in the doc/ directory in the future.