Remove outdated git_daemon_launched Windows info
This removes the Windows-specific information in the warning
message in git_daemon_launched.

After the associated functionality was updated in #1684 and the
warning message was abridged accordingly, the functionality was
updated again in #1697, causing the message to be outdated and no
longer helpeful (since having git-daemon.exe in a PATH directory is
no longer necessary or useful), without any corresponding change to
the message.