Pull HIDE_WINDOWS_*_ERRORS tests out of TestUtils
And make them pure pytest tests. This creates the TestEnvParsing
class, which does not inherit from TestCase.

Because unlike @ddt.data (and @ddt.idata) @pytest.mark.parametrize
can be nested (applied multiple times, creating tests for all
combinations), that is used for the HIDE_WINDOWS_*_ERRORS tests
instead of subtests. Because there were not any other uses of
subtests in the test suite, the pytest-subtest plugin is
accordingly removed from test-requirements. It may be put back
anytime in the future.