Move permission_error_tmpdir skips to test cases
This moves test skipping logic from the permission_error_tmpdir
fixture to the two TestRmtree test case methods that use it.

This produces some duplication, which is undesirable, but it makes
it so that which tests are skipped under what conditions is
immediately clear, easy to identify as skipping logic (which is
usually achieved by decoration), and clear in its relationship
to the skipping logic associated with other TestRmtree test cases.