Remove encoding declarations
In Python 3, UTF-8 is the default encoding when no encoding
declaration is present, and omitting the encoding declaration when
UTF-8 is intended is officially suggested.

- https://docs.python.org/3/reference/lexical_analysis.html
- https://pep8.org/#source-file-encoding

Furthermore, not all files in this project had them (and their
presence or absence seems to have been mainly for historical
reasons rather than file contents).