Use zero-argument super() where applicable
This replaces 2-argument super(...) calls with zero-argument
super() calls, where doing so preserves the exact semantics. This
turns out to be everywhere in the actual code (now all uses of the
super builtin are calls to it with zero arguments), and also in
most places code was discussed in comments.

The single exception, occurring in comments (technically a string
literal, but being used as a comment), appears in
git.objects.util.TraversableIterableObj, where super(Commit, self),
rather than super(TraversableIterableObj, self) which would have
the same meaning as super(), is shown. It may be that this should
be changed, but such a revision would not signify identical
semantics, and may require other changes to preserve or recover
clarity or documentary accuracy.