Make comments more consistent and clarify license
This improves the consistency of top-of-module comments as follows:

- All names of the current file are removed. Some included these
  while others didn't. In general, this can be useful information,
  which can remind readers and developers of what the file is and
  may even reduce mistakes. However, in GitPython, many modules
  inside git/ have the same name as other modules in other
  subdirectories of git/. So the presence of filenames would
  often be the same for multiple files, a condition that would be
  intensified if consistency were achieved by adding them
  everywhere. This instead removes them, which should (albeit
  slightly) decrease the risk of confusing modules that have the
  same name as each other.

- All modules (.py files) inside git/ and test/, except for .py
  files that are entirely empty (without even comments) or are
  inside test/fixtures/, now have comments indicating the license
  and linking to it on opensource.org. Previously, some modules
  had this, while others did not.

The comment about the license is short, and does not contain an
explicit copyright statement. No new explicit copyright statements
are added, but some top-of-modules already contained them, and this
does not remove (nor update or otherwise modify) them.

Although explicit copyright statements are not touched, all the
license comments are modified, including where they had previously
appeared, to say "the 3-Clause BSD License" instead of
"the BSD License", since there is no specific license known as the
"BSD License" (and both the 2-clause and 3-clause BSD licenses are
very popular).

This change should not be confused with #1662, which fixed an
originally correct hyperlink that had come to redirect to a page
about a different license. The change here does not change the link
again. It makes the commented wording more specific, so that it is
clear, even without looking at the link, which BSD license is being
referred to.