Update README to clarify license name
The license in README.md was named as "New BSD License", which is
in practice unambiguous, but no longer as readily recognized as is
"3-Clause BSD License". This updates the wording to call it by the
latter name, for that reason and for consistency with the previous
commit's changes to top-of-module comments, while also noting
parenthetically that it is also called the New BSD License.

The main reason for retaining "New BSD License" parenthetically is
to prevent anyone from being confused into thinking this change is
in any way associated with a change to the license, or that any
license change has occurred or is forthcoming. (This does *not*
change how GitPython is actually licensed in any way.)