Pass more specific license argument in setup.py
This uses "BSD-3-Clause" instead of "BSD" as the "license" metadata
in setup.py. "BSD-3-Clause" is the SPDX license identifier for the
BSD 3-Clause "New" or "Revised" License (the license GitPython
uses): https://spdx.org/licenses/BSD-3-Clause.html

There is no requirement to use an SPDX license identifier here,
but it is one of the common approaches, and it has the advantage of
making unambiguously clear, when a package is published on PyPI,
exactly what license it uses.

In contrast, the license-related *classifier* is unchanged, since
no more specific classfifier than what is in use now is currently
available.

The combination should result in

    License: BSD License (BSD-3-Clause)

being shown under "Meta" on PyPI, as of the next PyPI release. This
can be seen in other projects that use this combination of license
keyword argument and license-related classifier, such as:
https://pypi.org/project/flask-restx/