Extract some shared patching code to a helper
Because of how it is parameterized, but only in some of the test
cases that use it, making this a pytest fixture would be cumbersome
without further refactoring, so this is a helper method instead.

(It may be feasible to further refactor the test suite to improve
this more.)

This also removes a type annotation from a test method. It may
eventually be helpful to add annotations, in which case that would
come back along with others (if that test still exists in this
form), but it's the only test with such an annotation in this
test class, and the annotation had been added accidentally.