Avoid subprocess-writable temp file race condition
This lets the Windows subprocess open or rename onto the temporary
file using a more robust approach than in b12fd4a, avoiding the
race condition described there, where the filename could be
inadvertently reused between deletion and recreation of the file.

This creates a context manager helper for the temporary index file
used in IndexFile.from_tree, whose implementation differs by
operating system:

- Delegating straightforwardly to NamedTempoaryFile on POSIX
  systems where an open file can replaced by having another file
  renamed to it (just as it can be deleted).

- Employing custom logic on Windows, using mkstemp, closing the
  temporary file without immediately deleting it (so it won't be
  reused by any process seeking to create a temporary file), and
  then deleting it on context manager exit.

IndexFile.from_tree now calls this helper instead of
NamedTemporaryFile. For convenience, the helper provides the path,
i.e. the "name", when entered, so tmp_index is now just that path.

(At least for now, this helper is implemented as a nonpublic
function in the git.index.base module, rather than in the
git.index.util module. If it were public in git.index.util like the
other facilities there, then some later changes to it, or its later
removal, would be a breaking change to GitPython. If it were
nonpublic in git.index.util, then this would not be a concern, but
it would be unintuitive for it to be accessed from code in the
git.index.base module. In the future, one way to address this might
be to have one or more nonpublic _util modules with public members.
Because it would still be a breaking change to drop existing public
util modules, that would be more utility modules in total, so such
a change isn't included here just for this one used-once function.)