Add xfail marks for IndexFile.from_tree failures
8 of the tests that fail on native Windows systems fail due to
IndexFile.from_tree being broken on Windows, causing #1630.

This commit marks those tests as xfail. This is part, though not
all, of the changes to get CI test jobs for native Windows that
are passing, to guard against new regressions and to allow the code
and tests to be gradually fixed (see discussion in #1654).

When fixing the bug, this commit can be reverted.